+++ Markology/NoteController.swift
@@ -1,8 +1,10 @@
 import Combine
+import GRDBPlus
 import MarkCell
 import Markdown
 import MarkView
 import Notes
+import NotesUI
 import Paths
 import UIKit
 import UIKitPlus
@@ -48,70 +50,14 @@ class NoteController: UIViewController, Bindable {
         return tableView
     }()
 
-    private var linkSink: AnyCancellable?
-    private var linkQuery: ID.Search? {
-        didSet {
-            guard let linkQuery = linkQuery else {
-                linkSink = nil
-                return
-            }
-            linkSink = Engine.subscribe(with(\.link), to: linkQuery)
-        }
-    }
-
-    private var showLinks: Bool = false {
-        didSet {
-            collectionView.isHidden = !showLinks
-            let offset = showLinks ? 50.0 : 0
-            tableView.contentInset.bottom = offset
-            tableView.verticalScrollIndicatorInsets.bottom = offset
-        }
-    }
-
-    private var link: [ID] = [] {
-        didSet {
-            showLinks = true
-            collectionView.reloadData()
-            collectionView.setContentOffset(.zero, animated: false)
-        }
-    }
-
-    var search: String? {
-        didSet {
-            guard search != oldValue else { return }
-            guard let search = search else {
-                showLinks = false
-                linkQuery = nil
-                return
-            }
-            linkQuery = ID.search(text: search, limit: 5)
-        }
-    }
-
-    lazy var collectionView: UICollectionView = {
-        let size = 40.0
-        let layout = UICollectionViewFlowLayout()
-        layout.scrollDirection = .horizontal
-        layout.estimatedItemSize = CGSize(width: size, height: size)
-        layout.headerReferenceSize = CGSize(width: size, height: size)
-        layout.sectionFootersPinToVisibleBounds = true
-        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout).pinned(toKeyboardAnd: view, top: false)
-        collectionView.register(LinkCell.self)
-        collectionView.register(header: Header.self)
-        collectionView.backgroundColor = .clear
-        collectionView.dataSource = self
-        collectionView.delegate = self
-        collectionView.heightAnchor.constraint(equalToConstant: size).isActive = true
-        collectionView.isHidden = true
-        return collectionView
-    }()
+    let linkController = LinkController()
 
     private lazy var menuButton = UIBarButtonItem(image: UIImage(systemName: "ellipsis"), style: .plain, target: self, action: #selector(menu))
 
     var edit = false
     @objc func toggleEdit() {
         if edit {
-            search = nil
+            linkController.search = nil
             sync()
         }
         edit = !edit
@@ -134,15 +80,13 @@ class NoteController: UIViewController, Bindable {
                 }
                 return
             }
-            entrySink = Engine.subscribe(with(\.entry), to: Entry.load(id: id))
+            entrySink = Engine.shared.subscribe(with(\.entry), to: Entry.load(id: id))
         }
     }
 
     var text: String {
         document?.text ?? entry?.text ?? ""
     }
-
-    var addLink = PassthroughSubject<(url: String, text: String), Never>()
 
     func navigate(to id: ID) {
         guard let nav = navigationController else { return }
@@ -216,7 +160,8 @@ class NoteController: UIViewController, Bindable {
         super.viewDidLoad()
         view.backgroundColor = .systemBackground
         tableView.topAnchor.constraint(equalTo: view.topAnchor).isActive = true
-        collectionView.backgroundView = UIVisualEffectView(effect: UIBlurEffect(style: .systemChromeMaterial))
+        linkController.delegate = self
+        add(linkController)
         loaded = true
     }
 
@@ -325,6 +270,17 @@ class NoteController: UIViewController, Bindable {
     }
 }
 
+extension NoteController: LinkControllerDelegate {
+    func subscribe<T>(_ action: @escaping (T.Value) -> Void, to query: T) -> AnyCancellable where T: GRDBPlus.Query, T.Value: Equatable {
+        Engine.shared.subscribe(action, to: query)
+    }
+
+    func adjustInset(by offset: CGFloat) {
+        tableView.contentInset.bottom = offset
+        tableView.verticalScrollIndicatorInsets.bottom = offset
+    }
+}
+
 extension NoteController: MarkCellDelegate, EditCellDelegate {
     func change(text: String) {
         guard let document = document else { return }
@@ -342,12 +298,6 @@ extension NoteController: MarkCellDelegate, EditCellDelegate {
     func resolve(path: String) -> String? {
         guard let relative = id?.file.use(forEncoded: path) else { return nil }
         return relative.url.path.removingPercentEncoding
-    }
-}
-
-extension NoteController: SearchDelegate {
-    func change(search: String) {
-        self.search = search
     }
 }
 
@@ -412,7 +362,7 @@ extension NoteController: UITableViewDataSource {
             }
             return tableView.render((text: text, with: self), for: indexPath) as MarkCell
         case .edit:
-            return tableView.render((text: text, with: self), for: indexPath) as EditCell
+            return tableView.render((text: text, with: self, search: linkController), for: indexPath) as EditCell
         case .from:
             return tableView.render((link: entry!.from[indexPath.row], note: entry!.name), for: indexPath) as Entry.Link.Cell
         case .to:
@@ -448,94 +398,6 @@ extension NoteController: UITableViewDelegate {
             return
         }
         navigate(to: dest)
-    }
-}
-
-extension NoteController: UICollectionViewDataSource {
-    var linkSections: [MenuController.Section] {
-        var sections: [MenuController.Section] = []
-        if link.count > 0 {
-            sections.append(.notes)
-        }
-        sections.append(.new)
-        return sections
-    }
-
-    func numberOfSections(in collectionView: UICollectionView) -> Int {
-        linkSections.count
-    }
-
-    func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
-        switch linkSections[indexPath.section] {
-        case .notes:
-            return collectionView.render(linkQuery.valid ? "clock" : "magnifyingglass", forHeader: indexPath) as Header
-        case .new:
-            return collectionView.render("plus", forHeader: indexPath) as Header
-        }
-    }
-
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        switch linkSections[section] {
-        case .notes:
-            return link.count
-        case .new:
-            return 1
-        }
-    }
-
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        switch linkSections[indexPath.section] {
-        case .notes:
-            return collectionView.render(link[indexPath.row].name, for: indexPath) as LinkCell
-        case .new:
-            return collectionView.render(search ?? "", for: indexPath) as LinkCell
-        }
-    }
-}
-
-extension NoteController: UICollectionViewDelegate {
-    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        let target: ID
-        switch linkSections[indexPath.section] {
-        case .notes:
-            target = link[indexPath.row]
-        case .new:
-            guard let search = search else { return }
-            target = .generate(for: search)
-        }
-        guard let url = target.file.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed) else { return }
-        addLink.send((url: url, text: target.name))
-    }
-}
-
-extension NoteController {
-    class LinkCell: UICollectionViewCell, RenderCell {
-        lazy var label: UILabel = {
-            UILabel().pinned(to: contentView, layout: true)
-        }()
-
-        func render(_ text: String) {
-            if text == "" {
-                label.text = "Empty Note"
-                label.textColor = .placeholderText
-            } else {
-                label.text = text
-                label.textColor = .label
-            }
-        }
-    }
-}
-
-extension NoteController {
-    class Header: UICollectionReusableView, RenderCell {
-        lazy var image: UIImageView = {
-            UIImageView().pinned(to: self, withInset: 10)
-        }()
-
-        func render(_ symbol: String) {
-            image.image = UIImage(systemName: symbol)
-            image.tintColor = .secondaryLabel
-        }
     }
 }
